requirements.txt
setuptools>=65.0.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
distro==1.9.0
Django==5.2.4
django-appconf==1.1.0
django-compressor==4.5.1
django-fernet-fields==0.6
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
django-filter==25.1
drf-spectacular==0.28.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
nexus-rpc==1.1.0
openai==1.97.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
rcssmin==1.1.2
requests==2.32.4
rjsmin==1.2.2
sniffio==1.3.1
sqlparse==0.5.3
temporalio==1.14.1
tqdm==4.67.1
types-protobuf==6.30.2.20250703
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
# Redis для кеширования
redis==5.2.0
django-redis==5.4.0
django-ratelimit==4.1.0